requirements.txt
fastapi
uvicorn
google-cloud-speech
google-cloud-translate
google-cloud-texttospeech
pydub
numpy
